Add cmake stuff and update launcher
Modifies some of the CMakeLists.txt's to make one huge solution when just doing "cmake <rg dir>".

Adds batch file for default build in msvcbuild.

Refactors launcher and adds new logging system and a synchronization system for downloading only parts of files when updating.

Adds a "PatchCreator" application to create patch data server-side. Replaces makepatch.py.

Misc. stuff:
- Extend optional with nullopt_t and value constructor
- Add holds_alternative non-member function to variant.h for conformance with future std::variant
- Change some functions to return optionals
- Move reinterpret stuff from MUtil.h to new file reinterpret.h, and add more functions for disambiguating situations with pointers
- Move new code from Hash.h to new file MHash.h, remove Hash.h and Hash.c
- Remove unreferenced cml file list.c
- Refactor ArrayView
- Add MFile::File and MFile::RWFile
- Fix cml CMakeLists.txt not including .c files
- Remove MFile::FilesInDirRecursive because of implementation difficulty (there is no way to do it in constant memory)
- Move include/stdafx.h to stdafx.h in some project folders for consistency

+++ RealSpace2/Include/RMtrl.h
@@ -72,7 +72,7 @@ public:
 	int 	m_nAniTexCnt;
 	int 	m_nAniTexSpeed;
 	int 	m_nAniTexGap;
-	u32		m_backup_time;
+	u64		m_backup_time;
 
 	bool	m_bObjectMtrl;// effect ,interface , object
